[backend] split tests (#4538)
